Move #29286 tests to bundler_bun.test.ts; count sidecar cache hits
test/regression/issue is for true regressions; this guard never allowed
ESM bytecode without --compile, so the tests move next to the existing
target:bun bytecode coverage. The splitting test now requires two cache
hits (entry plus shared chunk) so a regression in the hashed-chunk
sidecar lookup cannot hide behind the entry's own hit.
